requirements.txt
streamlit
pillow
cryptography
pydub
